Add system that only changes titles
This is a preferred method to the one that set the `titleView` because
this works when pushing the new view controller and requires so much
less faith in doing the right thing with the content to display. The
title is restricted to only strings so the system will handle those as
it wants and SwiftUI is left to manage the `.principal` location toolbar
items however it wants.