Replace InterposeWrapper with simpler method
The original implementation used InterposeKit to hook into the life
cycle of the navigation bar. At the time I thought this was the only way
to get in ahead of the navigatin controller's push and pop of view
controllers to set the title and navigation item title. However, this
was more complicated than actually necessary. In fact, the system only
needed to have the navigation delegate be set to one that allows other
parts of the app to tell it what the next title should be.

Knowing what the next title should be is challening in SwiftUI because
you are supposed to write components that are ambivalent to when and
where they are displayed. The system provides the `navigationTitle` and
`navigationBarTitle` modifiers that allow you to set the title but those
have special functionality that ends up hooking into the wrapping view
controllers that SwiftUI creates. However, after some experimenting it
does seem that the order of evaluation of a View's `body` and any of its
children is pretty well defined, so really only the last evaluated
`body` needs to "win" in order to set the title as desired. That works
great and only requires a little modification to some code (maybe none
depending on style): when using a `NavigationLink` the destination
should not be defined inline but instead as a separate `View`.

Of course that all falls apart when pushing a `List` onto a
`NavigationView` because ... reasons. In that case no matter where the
new `navigationTitle(_:within:)` is declared the `View.body` is only
evaluated *after*
`UINavigationControllerDelegate.navigationController(willShow:animated:)`.
That is a problem. But further experimentation showed that if there was
an `AnyView` anywhere in the same `View.body` as the `List` that was
being pushed then SwiftUI would completely evaluate the `body` before
the push happened and so we would be back to where we wanted to be in
ordering. This was then accomplished by sticking one of the
`UIViewControllerRepresentable` instances in an `AnyView` because it is
at a leaf and should have little impact on the diffing perf of SwiftUI.